Handle word scoring with dynamic url better;
Changed the layout to add html elements if you are on a scoring page. Dynamic scoring with url or score/urlword can now handle more scoring using the score form after using dynamic url scoring. It also displays the most recently scored word along with its score correctly no matter if it was input through the url or the form.